moved mod to file and doc-commented
